rag-system/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
sqlmodel==0.0.14
alembic==1.12.1
aiopg==1.4.0
asyncpg==0.29.0

# Redis
redis==5.0.1

# 向量数据库
qdrant-client==1.6.9

# 对象存储
minio==7.2.0

# 环境配置
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP客户端
httpx==0.25.2

# 日志
loguru==0.7.2

# 工具库
click==8.1.7
typer==0.9.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.1.0
faker==20.1.0

# 文档解析
PyMuPDF==1.23.26
python-docx==1.1.2
chardet==5.2.0

# 自然语言处理
nltk==3.8.1
scikit-learn==1.3.2
langdetect==1.0.9

# 向量化和机器学习
sentence-transformers==2.2.2
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.21.0
numpy>=1.21.0
scipy>=1.7.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0